Allow some slippage while calculating machine steps in the evaluator
This allows us to do slightly inaccurate, but faster costing calculations.
The effect of this is that we can overshoot our budget by some amount.
We will eventually notice that we've gone over, and so
long as this is bounded reasonably, it seems worth it.

The way we do it is to:
1. Assume all the machine steps cost the same.
2. Keep a simple counter of the machine steps.
3. Only when the counter exceeds a threshold (the "slippage") do we do
properly account for the steps we've taken.

This isn't quite complete, since currently it has a fixed amount of
slippage (100 steps), which would e.g. be too high for very short
programs. We should probably instead do something like calculate it
based on a proportion of the total budget.

Overall this seems to give another 15-20% in benchmarks.